app_data_service/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-cors-headers==3.5.0
pika==1.1.0
pandas==1.3.4
numpy==1.21.4
yfinance==0.1.66
yahoo-fin==0.8.9.1
requests==2.26.0
python-dotenv==0.19.1